# src/types/users.ts
import { User } from "@prisma/client";

/**
 * Database use only
 */
type UserRecord = User;

type UserOutput = Omit<UserRecord, "is_deleted">;

type UserCreateInput = Pick<UserRecord, "email" | "fullname" | "country" | "user_type"> & { password: string };

type UserUpdateInput = Partial<Omit<UserRecord, "id">>;

export { UserRecord, UserOutput, UserCreateInput, UserUpdateInput };